Struct staticdir::StaticDir

pub struct StaticDir<T> {
    pub root: PathBuf,
    // some fields omitted
}

An implementation of Handler which serves list of directory contents.

Fields

root

The path from which this handler serves the list of directory contents.

Methods

impl<T> StaticDir<T>

fn new<P>(root: P, response_strategy: T) -> StaticDir<T> where P: Into<PathBuf>

Creates new instance of the StaticDir handler.