Trait staticdir::ResponseStrategy

pub trait ResponseStrategy {
    fn make_response(&self, dir: ReadDir) -> IronResult<Response>;
}

This trait is used by StaticDir to make a response from the collection of directory entries.

Required Methods

fn make_response(&self, dir: ReadDir) -> IronResult<Response>

Creates a response from the collection of directory entries.